requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cssselect==1.2.0
distro==1.9.0
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
lxml==5.3.0
MarkupSafe==3.0.2
nanoid==2.0.0
nh3==0.2.18
openai==1.54.4
pip-autoremove==0.10.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
requests==2.32.3
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
